move registerServiceWorker to services folder
